Decouple powerstrip & powerstrip-adapter from node
Previously we downloaded the binaries for powerstrip and
ran them inside the calico-node container. That makes the
individual components too tightly coupled, and hard to debug
individually.

Case in point, for a powerstrip bug I was asked for a minimum-
viable repro; sending over the whole calico-docker repo is not
feasible in this case, but building a stripped-down node-adapter
would be.